refactor: Add FaltasCuerpoTecnico resource to web.php
This commit adds the "FaltasCuerpoTecnico" resource to the web.php file. The resource is mapped to the "FaltasCuerpoTecnicoController" class, allowing for CRUD operations on the "faltasCuerpoTecnico" route. This addition enhances the functionality of the application by providing a new resource for managing "FaltasCuerpoTecnico" data.

Refactor routes/web.php to add "FaltasCuerpoTecnico" resource