refactor: Add id field to Equipos and FaltasCuerpoTecnico models
This commit adds the 'id' field to the Equipos model and FaltasCuerpoTecnico model. The 'id' field is now included in the fillable array of both models, allowing it to be mass assignable. This change ensures that the 'id' field can be set when creating or updating records in the database.

Refactor app/Models/Equipos.php and app/Models/FaltasCuerpoTecnico.php