refactor: Update FaltasCuerpoTecnicoController.php to join programaciones_faces table
This commit updates the FaltasCuerpoTecnicoController.php file to join the programaciones_faces table. The join condition is modified to include only the matching positions for the local and visitor teams. This change improves the accuracy of the query results when retrieving data related to the programaciones_faces table.

Refactor app/Http/Controllers/FaltasCuerpoTecnicoController.php to join programaciones_faces table