refactor: Update dependencies and fix UI issues
This commit updates the react-router-dom dependency to version 6.24.0 in the package.json file. It also fixes UI issues in the Index.jsx and Title.jsx files. The label in the ResultadosPartidos/Index.jsx file now includes the team name along with the player's name. In the Torneo/Title.jsx file, the styling of the title has been adjusted for better readability.

Refactor dependencies and fix UI issues