refactor: Add 'T.N.' option to cargo dropdown in CuerpoTecnico/Index.jsx
This commit adds the 'T.N.' (Tribuna) option to the cargo dropdown in the CuerpoTecnico/Index.jsx file. Now, users can select 'T.N.' as a valid option when adding or editing a member of the technical staff. This change ensures that the dropdown accurately reflects all available positions in the technical staff.

Refactor CuerpoTecnico/Index.jsx to include 'T.N.' option in cargo dropdown